+++ src/components/UnpublishedLabel/UnpublishedLabel.tsx
@@ -3,6 +3,7 @@ import * as React from 'react';
 import {block} from '../../utils';
 
 import {i18n} from './i18n';
+
 import './UnpublishedLabel.scss';
 
 const b = block('unpublished-label');